Move back to the version handling that we had before
