+++ cmd/fyne/internal/commands/deprecated.go
@@ -1,11 +1,8 @@
 package commands
 
 import (
-	"errors"
-	"fmt"
 	"os"
 	"os/exec"
-	"runtime/debug"
 
 	"github.com/urfave/cli/v2"
 )
@@ -21,24 +18,6 @@ func Vendor() *cli.Command {
 			cmd := exec.Command("go", "mod", "vendor")
 			cmd.Stdout, cmd.Stderr = os.Stdout, os.Stderr
 			return cmd.Run()
-		},
-	}
-}
-
-// Version returns the cli command for the program version.
-//
-// Deprecated: Use "fyne --version" or "fyne -v" instead.
-func Version() *cli.Command {
-	return &cli.Command{
-		Name:  "version",
-		Usage: "Deprecated: Use \"fyne --version\" or \"fyne -v\" instead.",
-		Action: func(_ *cli.Context) error {
-			if info, ok := debug.ReadBuildInfo(); ok {
-				fmt.Println("fyne cli version:", info.Main.Version)
-				return nil
-			}
-
-			return errors.New("could not retrieve version information (ensure module support is activated and build again)")
 		},
 	}
 }
