+++ cmd/fyne/internal/commands/version.go
@@ -1,0 +1,25 @@
+package commands
+
+import (
+	"errors"
+	"fmt"
+	"runtime/debug"
+
+	"github.com/urfave/cli/v2"
+)
+
+// Version returns the cli command for the program version.
+func Version() *cli.Command {
+	return &cli.Command{
+		Name:  "version",
+		Usage: "Shows version information for fyne.",
+		Action: func(_ *cli.Context) error {
+			if info, ok := debug.ReadBuildInfo(); ok {
+				fmt.Println("fyne cli version:", info.Main.Version)
+				return nil
+			}
+
+			return errors.New("could not retrieve version information (ensure module support is activated and build again)")
+		},
+	}
+}
